refactor(mvvm): co-locate view-models under lib/mvvm/viewmodels
The MVVM layer was split across two roots — screen controllers under
lib/mvvm/controllers/ but their view-models under lib/viewmodels/. Move
lib/viewmodels/* → lib/mvvm/viewmodels/* so the whole MVVM layer sits
under lib/mvvm/.

Pure structural move, no behaviour change:
- git mv the 12 view-models.
- Root-relative requires in controllers/device specs:
  require("viewmodels/X") → require("mvvm/viewmodels/X").
- Internal relative requires inside each VM go one level deeper:
  require("../util/X") → require("../../util/X") (sibling ./VM requires
  unchanged).
- Node specs: require("../../walta-app/app/lib/viewmodels/X") →
  .../lib/mvvm/viewmodels/X.
- Docs path references (viewmodels.md, modals.md, screen-controllers.md).

The Node spec dir test/viewmodels/ stays put — only the lib layer moves.
Verified: Node unit tests 632 passing; iOS device unit specs 83 passing
(runtime require resolution confirmed).
