+++ walta-app/app/lib/mvvm/viewmodels/SampleTrayPlus.js
@@ -1,4 +1,4 @@
-const ChangeNotifier = require("../util/ChangeNotifier");
+const ChangeNotifier = require("../../util/ChangeNotifier");
 
 const PLUS_ICON = "/images/plus-icon.png";
